chopin: power: Scan input devices for eligible DT2W node It is no longer a surprise for anyone to see Xiaomi shipping the same device with different displays and different displays usually mean different touchscreen drivers.
The current double tap handling in the power HAL only takes into
account one specific node, meant to be set in the device tree.
However, if the touchscreen was registered as a different input device
it would cause the feature to be broken. Address the issue by iterating
through the input devices in order to find one supported touchscreen
input to be used.

bengris32: Remove unused ts names for Xiaomi MT6768 platforms

Co-authored-by: LuK1337 <priv.luk@gmail.com>
Co-authored-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: Sushrut1101 <guptasushrut@gmail.com>